refactor(SearchTracePage): Co-locate search-panel-store with the component
The store is only consumed by SearchTracePage; keep it next to the
component instead of in the shared stores/ directory.
