+++ packages/jaeger-ui/src/components/SearchTracePage/search-panel-store.ts
